ref: Hoist `createCheckinEnvelope` to core package
