+++ packages/core/src/checkin.ts
