+++ packages/core/src/index.ts
@@ -17,6 +17,7 @@ export {
   setTags,
   setUser,
   withScope,
+  captureCheckIn,
 } from './exports';
 export {
   getCurrentHub,
@@ -42,6 +43,7 @@ export { SDK_VERSION } from './version';
 export { getIntegrationsToSetup } from './integration';
 export { FunctionToString, InboundFilters } from './integrations';
 export { prepareEvent } from './utils/prepareEvent';
+export { createCheckInEnvelope } from './checkin';
 export { hasTracingEnabled } from './utils/hasTracingEnabled';
 export { DEFAULT_ENVIRONMENT } from './constants';
