+++ packages/core/test/lib/checkin.test.ts
@@ -1,8 +1,8 @@
 import type { SerializedCheckIn } from '@sentry/types';
 
-import { createCheckInEnvelope } from '../src/checkin';
+import { createCheckInEnvelope } from '../../src/checkin';
 
-describe('CheckIn', () => {
+describe('createCheckInEnvelope', () => {
   test('creates a check in envelope header', () => {
     const envelope = createCheckInEnvelope(
       {
